integrate analysistools
